some clean up

+++ src/V3/engine_modules/moves.rs
@@ -749,34 +749,34 @@ impl Moves {
             let (r1, c1, r2, c2) = move_to_u32s!(move_str);
             let start_shift: u32 = 64 - 1 - (r1 * 8 + c1);
             let end_shift: u32 = 64 - 1 - (r2 * 8 + c2);
-            if ((1 << start_shift) & bitboards[Piece::WK]) != 0 { // white king move
+            if ((1i64 << start_shift) & bitboards[Piece::WK]) != 0 { // white king move
                 (castle_rights_t[CastleRights::CWK], castle_rights_t[CastleRights::CWQ]) = (false, false);
             }
-            if ((1 << start_shift) & bitboards[Piece::BK]) != 0 { // black king move
+            if ((1i64 << start_shift) & bitboards[Piece::BK]) != 0 { // black king move
                 (castle_rights_t[CastleRights::CBK], castle_rights_t[CastleRights::CBQ]) = (false, false);
             }
-            if ((1 << start_shift) & bitboards[Piece::WR] & 1) != 0 { // white king side rook move
+            if ((1i64 << start_shift) & bitboards[Piece::WR] & 1) != 0 { // white king side rook move
                 castle_rights_t[CastleRights::CWK] = false;
             }
-            if ((1 << start_shift) & bitboards[Piece::WR] & (1 << 7)) != 0 { // white queen side rook move
+            if ((1i64 << start_shift) & bitboards[Piece::WR] & (1 << 7)) != 0 { // white queen side rook move
                 castle_rights_t[CastleRights::CWQ] = false;
             }
-            if ((1 << start_shift) & bitboards[Piece::BR] & (1 << 56)) != 0 { // black king side rook move
+            if ((1i64 << start_shift) & bitboards[Piece::BR] & (1 << 56)) != 0 { // black king side rook move
                 castle_rights_t[CastleRights::CBK] = false;
             }
-            if ((1 << start_shift) & bitboards[Piece::BR] & (1 << 63)) != 0 { // black queen side rook move
+            if ((1i64 << start_shift) & bitboards[Piece::BR] & (1 << 63)) != 0 { // black queen side rook move
                 castle_rights_t[CastleRights::CBQ] = false;
             }
-            if (((1 as i64) << end_shift) & 1) != 0 { // white king side rook taken
+            if ((1i64 << end_shift) & 1) != 0 { // white king side rook taken
                 castle_rights_t[CastleRights::CWK] = false;
             }
-            if (((1 as i64) << end_shift) & (1 << 7)) != 0 { // white queen side rook taken
+            if ((1i64 << end_shift) & (1 << 7)) != 0 { // white queen side rook taken
                 castle_rights_t[CastleRights::CWQ] = false;
             }
-            if ((1 << end_shift) & ((1 as i64) << 56)) != 0 { // black king side rook taken
+            if ((1i64 << end_shift) & ((1 as i64) << 56)) != 0 { // black king side rook taken
                 castle_rights_t[CastleRights::CBK] = false;
             }
-            if ((1 << end_shift) & ((1 as i64) << 63)) != 0 { // black queen side rook taken
+            if ((1i64 << end_shift) & ((1 as i64) << 63)) != 0 { // black queen side rook taken
                 castle_rights_t[CastleRights::CBQ] = false;
             }
         }
